+++ src/main/scala/ai/evaluator/GeneralEvaluator.scala
@@ -4,6 +4,6 @@ import model.Board
 import model.Color.Color
 
 object GeneralEvaluator extends Evaluator {
-  override def Evaluate(board: Board, color: Color): Int = CheckmateEvaluator.Evaluate(board, color) +
+  override def Evaluate(board: Board, color: Color): Double = CheckmateEvaluator.Evaluate(board, color) +
     PieceScoreEvaluator.Evaluate(board, color)
 }
